for arras server
Bambi is the titular protagonist of Disney's 1942 animated feature film of the same name. A deer living in a forest, he is best friends with Thumper (a rabbit) and Flower (a skunk). However, he was closest to his mother, who was shot and killed by a hunter during his first winter. Raised into a buck by his father, he strongly falls in love with his childhood sweetheart and love interest, Faline, and they end up, later on, having twin fawns, a son and daughter named Geno and Gurri, respectively.
Bambi originated as the main character in Felix Salten's Bambi, a Life in the Woods. He has made cameos in several Disney cartoons. In the Disney films, his specie